pyproject.toml
[tool.poetry]
name = "exercise"
version = "0.1.0"
description = ""
authors = ["jhonatas"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
gunicorn = "^20.1.0"
prettyconf = "^2.2.1"
django-extensions = "^3.1.5"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest-reverse = "^1.5.0"
pytest-django = "^4.5.2"
pre-commit = "^2.19.0"
pytest = "^7.1.2"
factory-boy = "^3.2.1"
safety = "^1.10.3"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
